Refactored methods to separate file from models
