+++ server/index.js
@@ -1,13 +1,12 @@
 const express = require('express');
 const seq = require('../db/db.js');
-const Review = require('../db/Models/Review.js');
-const Characteristic = require('../db/Models/Characteristic.js');
-const Reviews_Photo = require('../db/Models/Reviews_Photo.js');
-const Characteristic_Review = require('../db/Models/Characteristic_Review.js');
+const Review_Api = require('../db/Reviews_Api.js');
 const app = express();
 const port = 3000;
 
-
+// Review.getChars(5).then((results) => {
+//   console.log(results);
+// });
 
 app.get('/', (req, res) => {
   res.send('Hello World!')
@@ -17,3 +16,8 @@ app.listen(port, () => {
   console.log(`Example app listening at http://localhost:${port}`)
 });
 
+// Review_Api.getCharReviewAverage(2).then((results) => {
+//   console.log(results);
+// }).catch((err) => {
+//   console.log(err);
+// });
